fix!: Do not export the filepicker Vue component in main entry point
Instead export it in a `filepicker` entrypoint, also load the component async. in the filepicker builder.
This ensures that if a user does not use the filepicker, it gets treeshaken.
